+++ com.unity.render-pipelines.high-definition/Editor/Lighting/HDLightUI.cs
@@ -699,7 +699,6 @@ namespace UnityEditor.Rendering.HighDefinition
 
             EditorGUI.BeginChangeCheck(); // For GI we need to detect any change on additional data and call SetLightDirty
 
-            // No cookie with area light (maybe in future textured area light ?)
             if (lightType != HDLightType.Area)
             {
                 serialized.settings.DrawCookie();
@@ -712,16 +711,50 @@ namespace UnityEditor.Rendering.HighDefinition
                     EditorGUILayout.PropertyField(serialized.shapeHeight, s_Styles.cookieSizeY);
                     EditorGUI.indentLevel--;
                 }
+
+                ShowCookieTextureTypeWarning(serialized.settings.cookie);
             }
             else if (serialized.areaLightShape == AreaLightShape.Rectangle)
             {
                 EditorGUILayout.ObjectField( serialized.areaLightCookie, s_Styles.areaLightCookie );
+                ShowCookieTextureTypeWarning(serialized.areaLightCookie.objectReferenceValue as Texture);
             }
 
             if (EditorGUI.EndChangeCheck())
             {
                 serialized.needUpdateAreaLightEmissiveMeshComponents = true;
                 SetLightsDirty(owner); // Should be apply only to parameter that's affect GI, but make the code cleaner
+            }
+        }
+
+        static void ShowCookieTextureTypeWarning(Texture cookie)
+        {
+            if (cookie == null)
+                return;
+
+            // The texture type is stored in the texture importer so we need to get it:
+            TextureImporter texImporter = AssetImporter.GetAtPath(AssetDatabase.GetAssetPath(cookie)) as TextureImporter;
+            
+            if (texImporter != null && texImporter.textureType == TextureImporterType.Cookie)
+            {
+                using (new EditorGUILayout.HorizontalScope())
+                {
+                    int indentSpace = (int)EditorGUI.IndentedRect(new Rect()).x;
+                    GUILayout.Space(indentSpace);
+                    using (new EditorGUILayout.HorizontalScope(EditorStyles.helpBox))
+                    {
+                        int oldIndentLevel = EditorGUI.indentLevel;
+                        EditorGUI.indentLevel = 0;
+                        GUIStyle wordWrap = new GUIStyle(EditorStyles.miniLabel){ wordWrap = true};
+                        EditorGUILayout.LabelField(s_Styles.cookieTextureTypeError, wordWrap);
+                        if (GUILayout.Button("Fix", GUILayout.ExpandHeight(true)))
+                        {
+                            texImporter.textureType = TextureImporterType.Default;
+                            texImporter.SaveAndReimport();
+                        }
+                        EditorGUI.indentLevel = oldIndentLevel;
+                    }
+                }
             }
         }